SCRUM-11 refactoring(services): User & Post service updated for backend connectivity
