+++ nexus-net/.env
@@ -1,0 +1,2 @@
+REACT_APP_USER_BASEURL=http://localhost:8080/
+REACT_APP_POST_BASEURL=http://localhost:8081/
